% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{cdf_gomp}
\alias{cdf_gomp}
\title{Compute the CDF of the Gompertz Distribution}
\usage{
cdf_gomp(x, alpha, beta)
}
\arguments{
\item{x}{A numeric vector of input values (failure times, for example) at which to evaluate the CDF.}

\item{alpha}{A numeric value representing the shape parameter of the Gompertz distribution. It controls the growth rate.}

\item{beta}{A numeric value representing the scale parameter of the Gompertz distribution. It affects the scale of the distribution.}
}
\value{
A numeric vector of the CDF values evaluated at the given points `x` using the specified parameters `alpha` and `beta`.
}
\description{
This function calculates the cumulative distribution function (CDF) of the Gompertz distribution for a given vector of input values, shape, and scale parameters.
}
\examples{
# Example CDF calculation
x <- c(1, 2, 3)  # Failure times
alpha <- 0.5     # Shape parameter
beta <- 0.1      # Scale parameter
cdf_gomp(x, alpha, beta)
}
