% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{compute_grad}
\alias{compute_grad}
\title{Compute the Numerical Gradient of the Log-Likelihood}
\usage{
compute_grad(Param)
}
\arguments{
\item{Param}{A numeric vector of parameters for which the gradient is calculated.}
}
\value{
A numeric vector of the same length as \code{Param}, representing the gradient at the specified parameters.
}
\description{
Calculates the gradient of the negative log-likelihood using finite differences.
The function uses a small step size (\code{h}) defined during initialization.
}
\details{
This function approximates the gradient using central finite differences.
Ensure that \code{h} is appropriately set to avoid numerical instability.
}
\examples{
\dontrun{
Param <- c(0.5, 0.1, 0.6, 0.2)
compute_grad(Param)
}
}
