% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{estimate_parameters_GCM}
\alias{estimate_parameters_GCM}
\title{Compute Cumulative Incidence Regression Results Generalized Chance Model (GCM)}
\usage{
estimate_parameters_GCM(initial_params, FeaturesNames = NULL)
}
\arguments{
\item{initial_params}{A numeric vector of initial parameter values to start the optimization. Default is \code{rep(0.001, 2 * (3 + ncol(covars)))}.}
}
\value{
A data frame containing:
\item{Params}{The parameter names ('alpha1', "tau1", 'rho1', 'alpha2', 'tau2', 'rho2', 'beta11', ..., 'beta1k', 'beta21', ..., 'beta2k').}
\item{S.E}{The standard deviations of the mean parameters.}
\item{Pvalue}{p-values.}
}
\description{
This function estimates the parameters of the model, computes the Hessian matrix, and calculates the variances and p-values for the parameters. It ensures that the diagonal elements of the covariance matrix are positive (GCM Approach).
}
\details{
This function performs the following steps:
\itemize{
\item Estimates the model parameters using the \code{optim} function and \code{log_f_rcpp} and \code{compute_log_f_gradient_rcpp}, \code{compute_log_f_hessian_rcpp}
\item Computes the Hessian matrix using the \code{hessian} function in \code{numDeriv} package.
\item Ensures that the diagonal elements of the covariance matrix are positive.
\item Calculates the variances and p-values for the parameters.
}
}
\examples{
\dontrun{
library(cmpp)
features <- matrix(rnorm(300, 1, 2), nrow = 100, ncol = 3)
delta1 <- sample(c(0, 1), 100, replace = TRUE)
delta2 <- 1 - delta1
x <- rexp(100, rate = 1/10)
Initialize(features, x, delta1, delta2, h = 1e-5)
initial_params <- rep(0.001, 2 * (ncol(features) + 3))
result <- estimate_parameters_GCM(initial_params)
print(result)
}

}
\seealso{
\link{optim}, \link{compute_log_f_gradient_rcpp}, \link{log_f_rcpp}, \link{compute_log_f_hessian_rcpp}.
}
