% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{GrLike}
\alias{GrLike}
\title{Compute the Gradient of the Log-Likelihood}
\usage{
GrLike(Param)
}
\arguments{
\item{Param}{A numeric vector of parameters (alpha, beta, etc.) for which the gradient should be computed.}
}
\value{
A numeric vector representing the gradient of the log-likelihood function at the specified parameters.
}
\description{
This function computes the gradient of the negative log-likelihood with respect to the model parameters using numerical differentiation.
}
\examples{
# Example gradient computation
Param <- c(0.5, 0.1, 0.6, 0.2)  # Example parameters
GrLike(Param)
}
