% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{Initialize}
\alias{Initialize}
\title{Initialize Data for the Cmpp Model}
\usage{
Initialize(features, x, delta1, delta2, h)
}
\arguments{
\item{features}{A matrix of features (predictor variables) for each observation. Each row corresponds to an observation, and each column corresponds to a feature.}

\item{x}{A vector of failure times corresponding to the observations. The failure times can represent the time to an event or censoring.}

\item{delta1}{A vector of Competing Risk indicator values. A binary vector where 1 indicates that the first event was observed and 0 means the event was censored or the second event occurred.}

\item{delta2}{A vector of Event indicator values. A binary vector where 1 indicates that the second event was observed, and 0 means the event was censored or the first event occurred.}

\item{h}{A small step size used for numerical gradient computation. It is essential for calculating the gradient in optimization routines.}
}
\value{
This function does not return any value. It initializes the Cmpp model with the provided data, which is then used in subsequent computations.
}
\description{
This function initializes the data used in the Cmpp model by passing the feature matrix, failure times, and the censoring indicators for competing risks and event indicators.
It sets up the necessary data structures for later computation.
}
\examples{
# Example data initialization for Cmpp model
features <- matrix(rnorm(100), ncol = 5)  # 20 observations with 5 features
x <- rnorm(20)  # Failure times for 20 observations
delta1 <- sample(0:1, 20, replace = TRUE)  # Competing risk indicator
delta2 <- sample(0:1, 20, replace = TRUE)  # Event indicator
Initialize(features, x, delta1, delta2, h = 1e-5)
}
