% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{LogLike1}
\alias{LogLike1}
\title{Compute the Log-Likelihood for the Model}
\usage{
LogLike1(param)
}
\arguments{
\item{param}{A numeric vector of model parameters: \code{alpha1}, \code{beta1}, \code{alpha2}, \code{beta2}, where
the first two are for the first event and the next two are for the second event.}
}
\value{
A single numeric value representing the negative log-likelihood.
}
\description{
Computes the negative log-likelihood of the Cmpp model given parameters and the initialized data.
The log-likelihood considers Gompertz distributions for competing risks.
}
\details{
This function requires the data to be initialized using \code{Initialize} before being called.
The log-likelihood is based on survival probabilities derived from the Gompertz distributions.
}
\examples{
\dontrun{
param <- c(0.01, 0.01, 0.01, 0.01)
LogLike1(param)
}
}
