% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCode.r
\name{pdf_gomp}
\alias{pdf_gomp}
\title{Compute the PDF of the Gompertz Distribution}
\usage{
pdf_gomp(x, alpha, beta)
}
\arguments{
\item{x}{A numeric vector of input values (failure times, for example) at which to evaluate the PDF.}

\item{alpha}{A numeric value representing the shape parameter of the Gompertz distribution.}

\item{beta}{A numeric value representing the scale parameter of the Gompertz distribution.}
}
\value{
A numeric vector of the PDF values evaluated at the given points `x` using the specified parameters `alpha` and `beta`.
}
\description{
This function calculates the probability density function (PDF) of the Gompertz distribution for a given vector of input values, shape, and scale parameters.
}
\examples{
# Example PDF calculation
x <- c(1, 2, 3)  # Failure times
alpha <- 0.5     # Shape parameter
beta <- 0.1      # Scale parameter
pdf_gomp(x, alpha, beta)
}
